+++ apis/compute/v1alpha1/datacenter_types.go
@@ -116,7 +116,7 @@ type DatacenterList struct {
 // Datacenter type metadata.
 var (
 	DatacenterKind             = reflect.TypeOf(Datacenter{}).Name()
-	DatacenterGroupKind        = schema.GroupKind{Group: Group, Kind: DatacenterKind}.String()
+	DatacenterGroupKind        = schema.GroupKind{Group: APIGroup, Kind: DatacenterKind}.String()
 	DatacenterKindAPIVersion   = DatacenterKind + "." + SchemeGroupVersion.String()
 	DatacenterGroupVersionKind = SchemeGroupVersion.WithKind(DatacenterKind)
 )