+++ cluster/local/integration_tests_compute.sh
@@ -977,3 +977,88 @@ EOF
   kubectl wait --for=delete ipfailovers/example
 
 }
+
+function group_tests(){
+
+  echo_step "deploy a group CR"
+  INSTALL_RESOURCE_YAML="$(
+    cat <<EOF
+apiVersion: compute.ionoscloud.crossplane.io/v1alpha1
+kind: Group
+metadata:
+  name: example
+managementPolicies:
+  - "*"
+spec:
+  forProvider:
+    name: exampleGroup
+    createDataCenter: true
+    reserveIp: true
+    createK8sCluster: true
+  providerConfigRef:
+    name: example
+EOF
+  )"
+
+  echo "${INSTALL_RESOURCE_YAML}" | "${KUBECTL}" apply -f -
+
+  echo_step "waiting for group CR to be ready & synced"
+  kubectl wait --for=condition=ready groups/example --timeout 120s
+  kubectl wait --for=condition=synced groups/example --timeout 120s
+
+  echo_step "get group CR"
+  kubectl get groups
+
+  echo_step "update group CR"
+  INSTALL_RESOURCE_YAML="$(
+    cat <<EOF
+apiVersion: compute.ionoscloud.crossplane.io/v1alpha1
+kind: Group
+metadata:
+  name: example
+managementPolicies:
+  - "*"
+spec:
+  forProvider:
+    name: exampleGroup
+    createDataCenter: true
+    reserveIp: true
+  providerConfigRef:
+    name: example
+EOF
+  )"
+
+  echo "${INSTALL_RESOURCE_YAML}" | "${KUBECTL}" apply -f -
+
+  echo_step "waiting for group CR to be ready & synced"
+  kubectl wait --for=condition=ready groups/example --timeout 120s
+  kubectl wait --for=condition=synced groups/example --timeout 120s
+
+}
+
+function group_tests_cleanup(){
+  INSTALL_RESOURCE_YAML="$(
+    cat <<EOF
+apiVersion: compute.ionoscloud.crossplane.io/v1alpha1
+kind: Group
+metadata:
+  name: example
+managementPolicies:
+  - "*"
+spec:
+  forProvider:
+    name: exampleGroup
+    createDataCenter: true
+    reserveIp: true
+  providerConfigRef:
+    name: example
+EOF
+  )"
+
+  echo_step "uninstalling group CR"
+  echo "${INSTALL_RESOURCE_YAML}" | "${KUBECTL}" delete -f -
+
+  echo_step "wait for deletion group CR"
+  kubectl wait --for=delete groups/example
+
+}